- (trivia) std::map 사용 코드 리팩토링
