+++ _MyLib_test.vcxproj
@@ -317,7 +317,7 @@ echo "done." </Command>
     <ClCompile Include="_test_rvo_move.cpp" />
     <ClCompile Include="_test_sched_client.cpp" />
     <ClCompile Include="_test_scm.cpp" />
-    <ClCompile Include="_test_std_map.cpp" />
+    <ClCompile Include="_test_std_map_set.cpp" />
     <ClCompile Include="_test_std_thread.cpp" />
     <ClCompile Include="_test_steady_timer.cpp" />
     <ClCompile Include="_test_template.cpp" />
